+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: pmedina- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/28 14:16:23 by pmedina-          #+#    #+#             */
-/*   Updated: 2020/10/29 13:03:20 by pmedina-         ###   ########.fr       */
+/*   Updated: 2020/11/10 13:53:31 by pmedina-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,9 +35,9 @@ void			printstring(t_printf *j, char *str);
 void			printint(t_printf *j, int num);
 void			printuns(t_printf *j, unsigned int num);
 size_t			intlen(long num);
-void			printhx(t_printf *j, size_t x);
-void			printhexu(t_printf *j, size_t x);
-void			printdir(t_printf *j, size_t dp);
+void			printhx(t_printf *j, unsigned long x);
+void			printhexu(t_printf *j, unsigned long x);
+void			printdir(t_printf *j, unsigned long dp);
 size_t			hxlen(size_t nb);
 
 #endif
